+++ electron/server-controller.js
@@ -19,8 +19,7 @@ function startServer(lnPath) {
     env: {
       PORT: 0 // any available port
     , LN_PATH: path.normalize(lnPath)
-    , ACCESS_KEY: accessKey
-    , LOGIN: `spark:${accessKey}`
+    , LOGIN: `spark:${accessKey}:${accessKey}`
     , NO_TLS: 1
     , NO_WEBUI: 1
     , NODE_ENV: 'production'