+++ src/auth.js
@@ -1,5 +1,7 @@
 import fs from 'fs'
+import path from 'path'
 import assert from 'assert'
+import mkdirp from 'mkdirp'
 import basicAuth from 'basic-auth'
 import nanoid from 'nanoid'
 import nanogen from 'nanoid/generate'
@@ -10,36 +12,38 @@ const cookieAge = 2592000000 // 1 month
 
 const hmacStr = (key, data) => createHmac('sha256', key).update(data).digest('base64').replace(/\W+/g, '')
 
-module.exports = (app, cookieFile, login, _accessKey) => {
-  if (cookieFile && (login || _accessKey)) throw new Error('--cookie-file cannot be used with --login or --access-key')
-
-  let username, password
+module.exports = (app, cookieFile, login) => {
+  let username, password, accessKey
 
   const hasFile = cookieFile && fs.existsSync(cookieFile)
-  if (hasFile) {
-    console.log(`Loading login credentials from ${cookieFile}`)
-    ;[ username, password, _accessKey ] = fs.readFileSync(cookieFile).toString('utf-8').trim().split(':')
-    assert(password, `Invalid login file at ${cookieFile}, expecting "username:pwd[:access-key]"`)
-  } else if (login) { // provided via --login (or LOGIN)
-    ;[ username, password ] = login.split(':', 2)
+
+  if (login) { // provided via --login (or LOGIN)
+    ;[ username, password, accessKey ] = login.split(':', 2)
     assert(password, `Invalid login format, expecting "username:pwd"`)
+  } else if (hasFile) {
+    console.log(`Loading login credentials from ${cookieFile}`)
+    ;[ username, password, accessKey ] = fs.readFileSync(cookieFile).toString('utf-8').trim().split(':')
+    assert(password, `Invalid login file at ${cookieFile}, expecting "username:pwd[:access-key]"`)
   } else { // generate random
     username = nanogen('abcdefghijklmnopqrstuvwxyz', 5)
     password = nanoid(15)
+    accessKey = hmacStr(`${username}:${password}`, 'access-key')
     console.log(`No LOGIN or --login specified, picked username "${username}" with password "${password}"`)
+
+    if (cookieFile) {
+      console.log(`Writing login credentials to ${cookieFile}`)
+      !fs.existsSync(path.dirname(cookieFile)) && mkdirp.sync(path.dirname(cookieFile))
+      fs.writeFileSync(cookieFile, [ username, password, accessKey ].join(':'))
+    }
   }
 
   // HMAC derive the access key from the user/pwd combo, if not explicitly defined
-  const accessKey   = _accessKey || hmacStr(`${username}:${password}`, 'access-key')
-      , manifestKey = hmacStr(accessKey, 'manifest-key').substr(0, 10)
+  accessKey || (accessKey = hmacStr(`${username}:${password}`, 'access-key'))
+
+  const manifestKey = hmacStr(accessKey, 'manifest-key').substr(0, 10)
       , manifestRe  = new RegExp(`^/manifest-${manifestKey}/`)
       , cookieKey   = hmacStr(accessKey, 'cookie-key')
       , cookieOpt   = { signed: true, httpOnly: true, sameSite: true, secure: app.enabled('tls'), maxAge: cookieAge }
-
-  if (cookieFile && !hasFile) {
-    console.log(`Writing login credentials to ${cookieFile}`)
-    fs.writeFileSync(cookieFile, [ username, password, accessKey ].join(':'))
-  }
 
   Object.assign(app.settings, { manifestKey, accessKey })