+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## Unreleased
+
+- Enable cookie file by default, persists random login credentials to `~/.spark-wallet/cookie`.
+
 ## 0.2.3 - 2019-01-31
 
 - docker: Update c-lightning to v0.6.3